TemplateImageOpenDialog: Move to separate files
